Udemy_Course_By_Krish_naike/GeminiApp/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain_community
groq
langchain-groq
pypdf
pdf2image